requirements.txt
flask
gunicorn
tensorflow-cpu
stripe
python-dotenv
firebase-admin
pillow
numpy